Create package.json
R: when using the automated-hardhat-boilerplate library as a starting point: illustrate example code descibing the following :

package.json: This file specifies the dependencies that your project needs in order to run. You will need to include any libraries or frameworks that you plan to use, such as web3.js for interacting with the Ethereum blockchain or bitcoind-promise for interacting with the Litecoin blockchain.

Chatgpt: Sure! Here is an example of what your package.json file might look like if you are using the automated-hardhat-boilerplate library as a starting point and need to include web3.js for interacting with the Ethereum blockchain and bitcoind-promise for interacting with the Litecoin blockchain:

{
  "name": "token-bridge",
  "version": "1.0.0",
  "main": "src/index.js",
  "license": "MIT",
  "dependencies": {
    "bitcoind-promise": "^4.0.0",
    "web3": "^2.6.0"
  }
}
This package.json file specifies two dependencies: bitcoind-promise and web3. These libraries will be installed when you run yarn in the project directory.

You can add additional dependencies to the dependencies object as needed. For example, if you also wanted to use the mongoose library to store transaction data in a MongoDB database, you could add it to the dependencies object like this:

"dependencies": {
  "bitcoind-promise": "^4.0.0",
  "web3": "^2.6.0",
  "mongoose": "^6.2.0"
}

I hope this helps! Let me know if you have any questions or need further clarification.