Create LitecoinToken.sol
R: when using the automated-hardhat-boilerplate library as a starting point: illustrate example code describing the following :

3. LitecoinToken.sol: This is the Solidity contract that defines the wLTC token on the Ethereum blockchain. It should implement the ERC20 token standard and define any additional functions or variables that you want your token to have. You can create this file in the contracts directory.

chatgpt: Sure! Here is an example of what your LitecoinToken.sol contract file might look like if you are using the automated-hardhat-boilerplate library as a starting point and want to define a token called wLTC that implements the ERC20 token standard:

pragma solidity ^0.8.0;

import "https://github.com/OpenZeppelin/openzeppelin-solidity/contracts/token/ERC20/SafeERC20.sol";

contract LitecoinToken is SafeERC20 {
  string public name = "wLTC";
  string public symbol = "wLTC";
  uint8 public decimals = 18;
  uint256 public totalSupply;

  constructor() public {
    totalSupply = 1000000 * (10 ** uint256(decimals));
  }
}

This LitecoinToken.sol contract defines a token called wLTC with a symbol of wLTC, 18 decimals, and a total supply of 1,000,000 tokens. It also imports the SafeERC20 contract from the OpenZeppelin library and inherits from it, which gives it the functions and events defined in the ERC20 standard.

You can customize this contract as needed to define any additional functions or variables that you want your token to have. For example, if you wanted to add a function that allows the owner of the contract to mint additional tokens, you could add a mint function like this:

function mint(address _to, uint256 _amount) public {
  require(msg.sender == owner, "Only the owner can mint tokens");
  require(_amount > 0, "Cannot mint 0 or negative tokens");
  totalSupply += _amount;
  balanceOf[_to] += _amount;
  emit Transfer(address(0), _to, _amount);
}

I hope this helps! Let me know if you have any questions or need further clarification.