Create index.js
R: when using the automated-hardhat-boilerplate library as a starting point: illustrate example code describing the following :

src/index.js: This is the main file for your bridging logic. It should contain the code that sets up a Litecoin node, listens for deposits of Litecoin on the Litecoin blockchain, and initiates the appropriate transactions on the Ethereum blockchain to mint the corresponding number of wLTC tokens.


chatgpt: Sure! Here is an example of what your src/index.js file might look like if you are using the automated-hardhat-boilerplate library as a starting point and want to set up a Litecoin node, listen for deposits of Litecoin on the Litecoin blockchain, and initiate the appropriate transactions on the Ethereum blockchain to mint the corresponding number of wLTC tokens:

const Web3 = require("web3");
const Bitcoin = require("bitcoind-promise");
const { fromWei } = require("web3-utils");

const web3 = new Web3(new Web3.providers.HttpProvider("http://localhost:8545"));
const litecoin = new Bitcoin({
  host: "localhost",
  port: 9332,
  user: "YOUR-LITECOIN-USERNAME",
  pass: "YOUR-LITECOIN-PASSWORD"
});

const LitecoinToken = new web3.eth.Contract(LITECOIN_TOKEN_ABI, LITECOIN_TOKEN_ADDRESS);

async function checkForDeposits() {
  const deposits = await litecoin.listTransactions("*", 100);
  for (const deposit of deposits) {
    if (deposit.category === "receive") {
      console.log(`Found deposit of ${deposit.amount} Litecoin!`);
      const balance = await litecoin.getBalance();
      console.log(`Current balance: ${balance} Litecoin`);
      if (balance >= deposit.amount) {
        const wLTCAmount = fromWei(deposit.amount.toString(), "ether");
        console.log(`Minting ${wLTCAmount} wLTC...`);
        await LitecoinToken.methods.mint(deposit.address, wLTCAmount).send({ from: OWNER_ADDRESS });
        console.log(`Successfully minted ${wLTCAmount} wLTC!`);
      } else {
        console.log("Insufficient balance to mint wLTC!");
      }
    }
  }
}

setInterval(checkForDeposits, 10000);

This src/index.js file uses the web3

R: continue previous output from this line : 
" This src/index.js file uses the web3 "

chatgpt: I apologize for the incomplete response earlier. Here is the rest of the example code for the src/index.js file:

[This src/index.js file uses the web3] library to set up a connection to the Ethereum blockchain and the `bitcoind-promise` library to set up a connection to the Litecoin blockchain. It defines a contract object for the `LitecoinToken` contract using the ABI and address of the deployed contract.

It then defines a `checkForDeposits` function that uses the `litecoin` object to get a list of the latest 100 transactions on the Litecoin blockchain and iterates over them. If it finds a transaction with a category of "receive", it logs a message and checks the current balance of the Litecoin node. If the balance is sufficient, it calls the `mint` function on the `LitecoinToken` contract and passes the address of the depositor and the amount of wLTC to mint as arguments.

Finally, it sets up an interval to call the `checkForDeposits` function every 10 seconds. This will continuously check for new deposits of Litecoin on the Litecoin blockchain and initiate the appropriate transactions on the Ethereum blockchain to mint the corresponding number of wLTC tokens.

I hope this helps! Let me know if you have any questions or need further clarification.