Create index.html
R: when using the automated-hardhat-boilerplate library as a starting point: illustrate example code describing the following :

src/index.html: If you are building a client-side application, you will need to create an HTML file that defines the user interface. This file should include any forms or input fields that users can use to initiate transfers of Litecoin to wLTC. Users should also be able to either deposit LTC and get wBTC or burn wBTC and receive LTC to a specified wallet address. Assume the ERC20 token contract implements the burn functions already.

chatgpt: This src/index.html file includes two forms: one for depositing Litecoin and receiving wLTC, and one for burning wLTC and receiving Litecoin. Each form includes an input field for the Ethereum address and an input field for the amount of Litecoin or wLTC to transfer, and a submit button to initiate the transfer.

I hope this helps! Let me know if you have any questions or need further clarification.